Build a responsive front end web pages for blog, blog detail with Angular and Bootstrap

+++ blog/src/main/java/com/morrun/blog/dao/BlogDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.morrun.blog.beans.Blog;
 
-public interface BlogDao extends JpaRepository<Blog, Integer> {
+public interface BlogDao extends JpaRepository<Blog, Long> {
 
 }
